+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Replace usage of the deprecated `qcodes.plots` with `qcodes_loop.plots` in the remote plot monitor (!465)
 - Visualization - Minor refactor to make visualization module compatible with `scipy>=1.11` (!466)
 - Deprecation - The `@deprecated` decorator now returns a function when decorating a function, instead of a class with a `__call__` method (!462).
+- Analysis - Added analysis class for qubit spectroscopy (!463)
 
 ## 0.7.2 (2023-05-02)